+++ client/src/app/details/submit/page.tsx
