+++ client/src/lib/useUser.ts
@@ -1,0 +1,19 @@
+import type { Profile } from "@/app/details/page";
+
+export default function useUser() {
+    const profile: Profile = {
+        firstNames: "Will",
+        lastNames: "Woodward",
+        preferredName: "Will",
+        pronouns: "He/Him",
+        age: "20",
+        phoneNumber: "+44 1234 567890",
+        email: "Will's email",
+        school: "Durham",
+        graduationYear: "2026",
+        levelOfStudy: "Level 3",
+        countryOfResidence: "United Kingdom"
+    }
+
+    return profile
+}
